+++ lib/src/models/models.dart
@@ -1,2 +1,2 @@
-export 'host_model.dart';
+export 'host.dart';
 export 'progress_callback.dart';